Revised jobs specification to support `cmake --build -jN`
Revised how jobs are specified.

For cmake 3.12+, we use `cmake --build -jN`.

For cmake < 3.12, we use `cmake --build [args] -- -jN -lN`, but only for Ninja and Makefiles generators.

Note we no longer check if we have a jobs based generator. This is now managed internally to `_get_jobs_arguments()`.

Also note we now add `--` inside `_get_jobs_arguments()` as it is not needed for CMake 3.12+