If available, use CMake file API to discover targets (#144)
The CMake file API was added in CMake 3.14. It can be used (among other
things) to get a list of targets for the project that was configured. In
my local tests, this appears to be significantly faster than crafting a
special invocation of the generator and should also work the same for
any generator.